requirements.txt
Django==4.2.15
django-environ==0.11.2
django-debug-toolbar==4.4.6
psycopg2-binary==2.9.9
djangorestframework==3.15.2
shortuuid==1.0.13
django-mptt==0.16.0
pillow==10.4.0
django-filter==24.3
redis==5.0.8
celery==5.4.0
pillow==10.4.0